Move formats away from public API
Formats for `AnalyzedClass`, `Compilation` and `Companions` were in the
public API (xsbti.api)! They have been removed from there.
