+++ internal/zinc-persist/src/main/scala/sbt/internal/inc/CompanionsFormat.scala
@@ -5,19 +5,18 @@
  * This software is released under the terms written in LICENSE.
  */
 
-package xsbt.api
+package sbt.internal.inc
 
-import xsbti.api._
 import sbinary._
-import sbt.internal.inc.Compilation
+import xsbti.api._
 
-object CompilationFormat extends Format[Compilation] {
+object CompanionsFormat extends Format[Companions] {
   import java.io._
-  def reads(in: Input): Compilation = {
+  def reads(in: Input): Companions = {
     val oin = new ObjectInputStream(new InputWrapperStream(in))
-    try { oin.readObject.asInstanceOf[Compilation] } finally { oin.close() }
+    try { oin.readObject.asInstanceOf[Companions] } finally { oin.close() }
   }
-  def writes(out: Output, src: Compilation): Unit = {
+  def writes(out: Output, src: Companions): Unit = {
     val oout = new ObjectOutputStream(new OutputWrapperStream(out))
     try { oout.writeObject(src) } finally { oout.close() }
   }
